requirements.txt
torch
torchvision
opencv-python
matplotlib
numpy
Pillow
ultralytics
streamlit